railstack-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.3.3
pydantic-settings==2.5.2
python-dotenv==1.0.1
